fix: stop animation when player pause his game
